+++ apps/pwa/tests/Overlays.spec.ts
@@ -9,18 +9,16 @@ test.describe('Loading images', () => {
 
     expect(await page.$$('canvas')).toHaveLength(1)
     expect(
-      await page.textContent('text=/matrix size: 207 x 256 x 215, voxelsize: 0.74 x 0.74 x 0.74/i'),
+      await page.textContent('text=/matrix size: .* voxelsize: .*/i'),
     ).toBeTruthy()
 
     await loadOverlay(page)
-
-    const message = listenForDebugMessage(page)
-
-    await page.evaluate(() => {
-      const message = { type: 'debugRequest', body: 'getNVolumes' }
-      window.postMessage(message, '*')
-    })
-
-    expect(await message).toBe(2)
+    await expect.poll(async () => {
+      const messagePromise = listenForDebugMessage(page)
+      await page.evaluate(() => {
+        window.postMessage({ type: 'debugRequest', body: 'getNVolumes' }, '*')
+      })
+      return await messagePromise
+    }, { timeout: 10000 }).toBe(2)
   })
 })